Split types and declarations into separate files
