+++ common/decls/BUILD.bazel
@@ -9,10 +9,26 @@ go_library(
     name = "go_default_library",
     srcs = [
         "decls.go",
+        "types.go",
     ],
     importpath = "github.com/google/cel-go/common/decls",
     deps = [
         "//common/functions:go_default_library",
+        "//common/types:go_default_library",
+        "//common/types/ref:go_default_library",
+        "//common/types/traits:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "decls_test.go",
+        "types_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = [
+        "//common/overloads:go_default_library",
         "//common/types:go_default_library",
         "//common/types/ref:go_default_library",
         "//common/types/traits:go_default_library",
